fix: send the gem index as U16 in wheel gem actions
Canary reads a uint16_t for Destroy, SwitchDomain and ToggleLock in the
0xE7 wheel packet, but the client sent the gem index as U8 - the server
read 0 in silence and acted on the gem at index 0 (destroying or locking
the wrong gem).

+++ src/client/game.cpp
@@ -2338,7 +2338,7 @@ void Game::openWheel(uint32_t playerId)
     m_protocolGame->sendOpenWheel(playerId);
 }
 
-void Game::gemAction(const uint8_t actionType, const uint8_t param, const uint8_t pos)
+void Game::gemAction(const uint8_t actionType, const uint16_t param, const uint8_t pos)
 {
     if (!canPerformGameAction())
         return;
